refactor: kafka 스키마 파일 위치 변경 및 import 경로 수정
